requirements.txt
python-chess
torch
requests
chess-tournament
huggingface-hub
numpy